+++ app/controllers/meets_controller.rb
@@ -1,6 +1,7 @@
 class MeetsController < ApplicationController
     before_action :set_availability_tutor, only: [ :index ]
-    before_action :set_meet, only: [ :show ]
+    before_action :set_meet, only: [ :show, :confirm_pending_meet ]
+    before_action :set_date, only: [ :confirm_pending_meet ]
 
     # GET /availability_tutors/:availability_tutor_id/meets
     def index
@@ -13,6 +14,21 @@ class MeetsController < ApplicationController
     def show
       # Return a specific meeting's details
       render json: @meet, status: :ok
+    end
+
+    # POST /meets/:id
+    def confirm_pending_meet 
+      @availability_tutor = AvailabilityTutor.find(@meet.availability_tutor_id)
+      if @current_user.first.id != @availability_tutor.user_id
+        render json: { error: "User not allowed" }, status: :unauthorized
+      elsif @meet.status != "pending"
+        render json: { error: "Meet already confirmed" }, status: :bad_request
+      else
+        @meet.status = "confirmed"
+        @meet.date_time = params[:meet][:date]
+        @meet.save
+        render json: { message: "Meet confirmed successfully" }, status: :ok
+      end
     end
 
     private
@@ -29,5 +45,10 @@ class MeetsController < ApplicationController
       @meet = Meet.find(params[:id])
     rescue ActiveRecord::RecordNotFound
       render json: { error: "Meet not found" }, status: :not_found
+    end
+
+    # Find the date in timestamp in the URL
+    def set_date
+      @dateTS = params.require(:meet).permit[:date]
     end
 end